+++ src/transformers/generation/continuous_batching/__init__.py
@@ -13,7 +13,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 from .cache import PagedAttentionCache
-from .classes import RequestState, RequestStatus
+from .requests import RequestState, RequestStatus
 from .continuous_api import ContinuousBatchingManager, ContinuousMixin
